+++ MANIFEST.in
@@ -1,2 +1,2 @@
 include LICENCE *.rst
-recursive-include kinto2xml *.json *.xml
+recursive-include amo2kinto *.json *.xml
