+++ amo2kinto/compare.py
