+++ amo2kinto/constants.py
