+++ amo2kinto/kinto.py
