+++ amo2kinto/logger.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger("amo2kinto")
