+++ amo2kinto/synchronize.py
