+++ amo2kinto/tests/fixtures/blocklist.xml
