+++ amo2kinto/tests/fixtures/blocklists.json
