+++ amo2kinto/tests/fixtures/fennec-blocklist.xml
