+++ amo2kinto/tests/fixtures/generated-blocklist.xml
