+++ amo2kinto/verifier.py
