+++ tox.ini
@@ -5,7 +5,7 @@ skip_missing_interpreters = True
 [testenv]
 commands =
     python --version
-    py.test --cov-report term-missing --cov-fail-under 100 --cov kinto2xml {posargs} kinto2xml
+    py.test --cov-report term-missing --cov-fail-under 100 --cov amo2kinto {posargs} amo2kinto
 deps = -rdev-requirements.txt
 install_command = pip install --pre {opts} {packages}
 
@@ -15,6 +15,6 @@ commands =
     kinto2xml -s http://localhost:8888/v1
 
 [testenv:flake8]
-commands = flake8 kinto2xml
+commands = flake8 amo2kinto
 deps =
     flake8
